Cache the compiled diffusion model so we don't have to recompile every R session.
Unfortunately it doesn't appear safe to use this cache to load the compiled model on each Rsession in a cluster, trying to do so results in random "Evaluation error: zero byte input" errors. Probably some race condition in the writing the .R file that gets sourced in to load the Rcpp functions.